requirements.txt
alpaca-py>=0.21.0
python-dotenv>=1.0.0
yfinance>=1.2.0
pandas>=2.2.0
numpy>=1.26.0
jinja2>=3.1.0
requests>=2.31.0
google-generativeai>=0.8.0
pandas-datareader>=0.10.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
